+++ central/clusters/identity_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/stackrox/rox/generated/storage"
-	testutilsMTLS "github.com/stackrox/rox/pkg/testutils/central/mtls"
+	testutilsMTLS "github.com/stackrox/rox/pkg/mtls/testutils"
 	"github.com/stackrox/rox/pkg/testutils/envisolator"
 	"github.com/stretchr/testify/suite"
 )
