+++ central/localscanner/certificates_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stackrox/rox/pkg/certgen"
 	"github.com/stackrox/rox/pkg/features"
 	"github.com/stackrox/rox/pkg/mtls"
-	testutilsMTLS "github.com/stackrox/rox/pkg/testutils/central/mtls"
+	testutilsMTLS "github.com/stackrox/rox/pkg/mtls/testutils"
 	"github.com/stackrox/rox/pkg/testutils/envisolator"
 	"github.com/stretchr/testify/suite"
 )
