+++ central/sensor/service/connection/connection_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stackrox/rox/pkg/centralsensor"
 	"github.com/stackrox/rox/pkg/concurrency"
 	"github.com/stackrox/rox/pkg/features"
-	testutilsMTLS "github.com/stackrox/rox/pkg/testutils/central/mtls"
+	testutilsMTLS "github.com/stackrox/rox/pkg/mtls/testutils"
 	"github.com/stackrox/rox/pkg/testutils/envisolator"
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/grpc"
