+++ pkg/mtls/testutils/testdata/central-certs/ca-key.pem
